Start to organize AHCI, also add qemu logging using COM1 serial port

+++ src/include/assembly.h
@@ -31,9 +31,9 @@ void switch_stack(void* new_stack, void (*func)(void));
 
 void load_gdtr(void* gdtr);
 
-uint8_t port_inb(uint8_t port);
+uint8_t port_inb(uint16_t port);
 
-void port_outb(uint8_t port, uint8_t data);
+void port_outb(uint16_t port, uint8_t data);
 
 port_outl(uint16_t port, uint32_t val);
